sleep na ako

+++ patient/deleteprescription.php
@@ -12,8 +12,23 @@ if (!isset($_SESSION['patientSession'])) {
 if (isset($_POST['prescriptionId'])) {
     $prescriptionId = mysqli_real_escape_string($con, $_POST['prescriptionId']);
 
-    // Delete the prescription from the database
-    $deleteQuery = "DELETE FROM prenatalprescription WHERE prescriptionId = '$prescriptionId'";
+    // Check the appointment type of the prescription
+    $res = mysqli_query($con, "SELECT * FROM prenatalprescription WHERE prescriptionId = '$prescriptionId'");
+
+    if (!$res) {
+        // Error in the SQL query
+        die("Error checking prescription type: " . mysqli_error($con));
+    }
+
+    if (mysqli_num_rows($res) > 0) {
+        // Prescription is from the Prenatal table
+        $deleteQuery = "DELETE FROM prenatalprescription WHERE prescriptionId = '$prescriptionId'";
+    } else {
+        // Prescription is from the TB table
+        $deleteQuery = "DELETE FROM tbprescription WHERE prescriptionId = '$prescriptionId'";
+    }
+
+    // Perform the deletion
     $deleteResult = mysqli_query($con, $deleteQuery);
 
     if ($deleteResult) {
